% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{as_wactor}
\alias{as_wactor}
\title{As wactor}
\usage{
as_wactor(.x, ...)
}
\arguments{
\item{.x}{Input text vector}

\item{...}{Other args passed to Wactr$new(...)}
}
\value{
An object of type wactor
}
\description{
Convert data into object of type 'wactor'
}
