% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wactor-package.R
\docType{package}
\name{wactor-package}
\alias{wactor}
\alias{wactor-package}
\title{wactor: Word Factor Vectors}
\description{
A user-friendly factor-like interface for converting strings of 
    text into numeric vectors and rectangular data structures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mkearney/wactor}
  \item Report bugs at \url{https://github.com/mkearney/wactor/issues}
}

}
\author{
\strong{Maintainer}: Michael W. Kearney \email{kearneymw@missouri.edu} (0000-0002-0730-4694)

}
\keyword{internal}
