% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{wactor}
\alias{wactor}
\title{Create wactor}
\usage{
wactor(.x, ...)
}
\arguments{
\item{.x}{Input text vector}

\item{...}{Other args passed to Wactor$new(...)}
}
\value{
An object of type wactor
}
\description{
Create an object of type 'wactor'
}
