% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wactor-r6.R
\docType{data}
\name{Wactr}
\alias{Wactr}
\title{A wactor object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Wactr
}
\description{
A factor-like class for word vectors
}
\keyword{datasets}
