% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_mat.R
\name{xgb_mat}
\alias{xgb_mat}
\title{xgb matrix}
\usage{
xgb_mat(x, ..., y = NULL, split = NULL)
}
\arguments{
\item{x}{Input data}

\item{...}{Other data to cbind}

\item{y}{Label vector}

\item{split}{Optional number between 0-1 indicating the desired split between
train and test}
}
\value{
A xgb.Dmatrix
}
\description{
Simple wrapper for creating a xgboost matrix
}
\examples{

xgb_mat(data.frame(x = rnorm(20), y = rnorm(20)))

}
