% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wactor-s3.R
\name{dtm}
\alias{dtm}
\title{Document term frequency}
\usage{
dtm(object, .x = NULL)
}
\arguments{
\item{object}{Input object containing dictionary (column), e.g., wactor}

\item{.x}{Text from which the document term matrix will be created}
}
\value{
A c-style matrix
}
\description{
Converts character vector into document term matrix (dtm)
}
