% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wactor-s3.R
\name{wactor}
\alias{wactor}
\title{Create wactor}
\usage{
wactor(.x, ...)
}
\arguments{
\item{.x}{Input text vector}

\item{...}{Other args passed to Wactr$new(...)}
}
\value{
An object of type wactor
}
\description{
Create an object of type 'wactor'
}
\examples{

## create
w <- wactor(c("a", "a", "a", "b", "b", "c"))

## summarize
summary(w)

## plot
plot(w)

## predict
predict(w)

## use on NEW data
dtm(w, letters[1:5])

## dtm() is the same as predict()
predict(w, letters[1:5])

## works if you specify 'newdata' too
predict(w, newdata = letters[1:5])

}
